\name{attentiondata}
\alias{attentiondata}
\docType{data}
\title{
Demo dataset
}
\description{Time series from 'Attention to visual motion' dataset).

}
\usage{data(semdata)}
\format{
A data frame of 360 observations of 3 variables.
\describe{
    \item{\code{x}}{Time series of V1}
    \item{\code{y}}{Time series of V5}
    \item{\code{z}}{Time series of SPC}
  }
}
\examples{
head(attentiondata)
}

