\name{dcmCompare}
\alias{dcmCompare}
\title{DCM comparison
}
\description{Compute Bayes Factor based on fitvalues of 2 DCM's
}
\usage{
dcmCompare(DCM1, DCM2)
}
\arguments{
  \item{DCM1}{First DCM.}
  \item{DCM2}{Second DCM.}
}
\value{Returns a Bayes Factor based on both AIC's of the models and BIC's.
}
\author{Bjorn Roelstraete}
\seealso{spmEstimate}
\keyword{ ~DCM }
