\name{SEMextract}
\alias{SEMextract}
\title{Preparing fMRI time series for SEM analysis
}
\description{Preparing fMRI time series for SEM analysis.
}
\usage{
SEMextract(ts, ons, dur, TR)
}
\arguments{
  \item{ts}{time series for analysis
}
  \item{ons}{Onsets of experimental condition of interest.
}
  \item{dur}{Duration of experimental condition of interest.
}
\item{TR}{Repetition time of time series.
}
}
\author{Bjorn Roelstraete}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~SEM }

